@@ -1,11 +1,11 @@
 library;
 
 export 'core/audio_controller.dart';
-export 'core/media/song_metadata.dart';
+export 'core/model/song_metadata.dart';
 export 'core/media/song_metadata_util.dart';
-export 'core/playback/playback_track.dart';
+export 'core/model/playback_track.dart';
 export 'core/playback/playback_helper.dart';
-export 'core/playback_view_model.dart';
+export 'core/model/playback_view_model.dart';
 export 'core/library/library_source.dart';
 export 'core/library/library_playback_service.dart';
 export 'core/storage/library_storage.dart';
